requirements.txt
asgiref==3.5.2
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
djangorestframework==3.13.1
engineering-notation==0.6.0
fcm-django==1.0.12
firebase-admin==5.3.0
gcloud==0.18.3
google-api-core==2.10.0
google-api-python-client==2.58.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.1
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.48.1
grpcio-status==1.48.1
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
jws==0.1.3
msgpack==1.0.4
mysqlclient==2.1.1
oauth2client==4.1.3
Pillow==9.2.0
plyer==2.0.0
proto-plus==1.22.1
protobuf==4.21.5
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.15.0
PyJWT==2.4.0
pyparsing==3.0.9
python-jwt==2.0.1
pytz==2022.2.1
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
tk-tools==0.16.0
twilio==7.13.0
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
